packages/python-worker/requirements.txt
supabase==2.3.4
easyocr==1.7.1
deepface==0.0.93
python-dotenv==1.0.1
# Let DeepFace manage tensorflow, retina-face, opencv, and pillow
